♻️ refactor(contract): split write-DTO tests into ReadingOrderWritesTest
